+++ lib/features/home/presentation/live_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:just_audio/just_audio.dart';
-import 'package:youtube_player_flutter/youtube_player_flutter.dart';
+import 'package:youtube_player_flutter/youtube_player_flutter.dart' hide PlayerState;
 
 import '../../../core/services/radio_service.dart';
 import '../../../core/config/app_theme.dart';